+++ NEWS.md
@@ -1,6 +1,16 @@
 # simnetdemand
 
-# simnetdemand 1.0.0 2022_03_23
+## simnetdemand 1.0.1 2022_04_20
+
+* add `nhpp()` function
+* add `agg_bh()` function
+* add `forecast_demand()` function
+* add `forecast_demand_sim()` function
+* add `emsr_limits()` function
+* add `get_cumulative_counts()` function
+* add GitHub pages
+
+## simnetdemand 1.0.0 2022_03_23
 
 * initial release